refactor(decomp): extract cdc outbound → proximadb-cdc (TD-DECOMP-10)
Moves src/cdc/outbound/{config,dedup,exactly_once,mod,position,router,
subscriber} (57 tests) into the cdc crate (127→184 tests). Whole subdir
verbatim; rewrite crate::cdc::{event,error}→crate::{event,error}. Added
deps: uuid (inline 1.0/v4), regex (workspace). Root re-exports outbound.

Verified: cdc crate 184 tests pass; root check + server build + cdc-clippy
+ fmt + check_workspace_boundaries + check-layering all green.
