+++ crates/integrations/proximadb-cdc/Cargo.toml
@@ -29,3 +29,6 @@ tokio.workspace = true
 async-trait.workspace = true
 tracing.workspace = true
 reqwest = { version = "0.11", features = ["json"] }
+# outbound: uuid (position/event IDs), regex (event filtering).
+uuid = { version = "1.0", features = ["v4"] }
+regex.workspace = true
