+++ crates/integrations/proximadb-cdc/src/outbound/config.rs
