+++ docs/10-quality/td/TD-DECOMP-10-extract-cdc-outbound.adoc
@@ -1,0 +1,38 @@
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+= TD-DECOMP-10: Extract cdc outbound → proximadb-cdc (god-crate decomposition)
+
+[cols="1,3", options="header"]
+|===
+| Field | Value
+| Status | Landing (delivered by this slice)
+| Severity | Low — build/dev-loop only; behavior-neutral re-export
+| Component | `crates/integrations/proximadb-cdc` (`outbound`); root `src/cdc/mod.rs`
+| Relates to | [[TD-DECOMP-1]] (recipe); [[TD-DECOMP-7..9]] (cdc crate seed + transform/error/metrics + sinks)
+|===
+
+== Why
+
+Continues the cdc vein. The outbound subsystem (`{config,dedup,exactly_once,mod,position,
+router,subscriber}`, **57 tests**) is the next cohesive unit — its only non-self deps are
+`event` + `error` (both in the cdc crate). Deps: serde + async-trait + tracing + tokio +
+uuid + regex. The cdc crate goes 127→184 tests.
+
+== What moved (behavior-neutral)
+
+* `src/cdc/outbound/` (whole subdir) → `crates/integrations/proximadb-cdc/src/outbound/`.
+* Rewrite: `crate::cdc::{event,error}` → `crate::{event,error}`.
+* `proximadb-cdc/Cargo.toml`: + `uuid` (inline `1.0`/v4) + `regex` (workspace).
+* `proximadb-cdc/src/lib.rs`: `pub mod outbound;`.
+* Root `src/cdc/mod.rs`: `pub mod outbound;` → `pub use proximadb_cdc::outbound;`.
+
+== Verification
+
+* `cargo check -p proximadb-cdc` green; `nextest` — 184 tests pass.
+* `cargo check -p proximadb` + `check_workspace_boundaries` + `check-layering` + server
+build + cdc-clippy + fmt + work-commit-check green.
+
+== Follow-ups
+
+* `connectors/{mongodb,mysql,postgres}/*` (~80 tests) + `coordinator` (the hub, ~11 tests).
+* **Running total once merged: ~309 tests out of root lib binary** (252 + 57).
